pyproject.toml
[tool.poetry]
name = "citations_searcher"
# Version changes automatically
version = "1.0.0"
description = "Searches for citations of other papers in yours."

authors = [
    "Sergei Zakharov <thesdzakharov@gmail.com>"
]
license = "MIT License"

###############################################################################
[tool.poetry.dependencies]
python = '^3.10'
pandas = "^2.2.1"
tqdm = "^4.66.2"
loguru = "^0.7.2"
sqlalchemy = "^2.0.29"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
catboost = "^1.2.5"
scikit-learn = "^1.4.2"
torch = "^2.3.0"
transformers = "^4.40.2"
wandb = "^0.17.0"
django = "^5.0.6"
celery = "^5.4.0"
redis = "^5.0.4"
hydra-core = "^1.3.2"

[tool.poetry.group.dev.dependencies]
pytest = "*"
jupyter = "^1.0.0"

# Запускаем coverage тестов только на папку src/
[tool.coverage.run]
source = ['citations_searcher']

###############################################################################

[tool.isort]
line_length = 120
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
no_lines_before = ["STDLIB", "LOCALFOLDER"]
known_third_party = []
multi_line_output = 3
lines_after_imports = 2
profile = "black"
combine_as_imports = true

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.ruff]
fix = true

lint.select = ["ALL"]
lint.ignore = ["I","EM","FBT","TRY003","S101","D101","D102","D103","D104","D105","G004","D107","D211","D213","FA102","ANN101", "FIX002", "TD002", "TD003"]

lint.fixable = ["ALL"]
lint.unfixable = ["F841"]

exclude = [
    ".git",
    ".ruff_cache",
    "venv",
]

line-length = 120

target-version = "py310"

cache-dir = "~/.cache/ruff"

[tool.ruff.lint.mccabe]
max-complexity = 10


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
